+++ geoips/commandline/test_interfaces.py
@@ -20,21 +20,19 @@ from importlib import import_module
 import traceback
 
 
-def main():
-    """Script to test all dev and stable interfaces."""
-    # Removed all interfaces that have been moved to new setup.
-    # This entire testing construct will be updated in the future to more fully
-    # test/validate the plugins.
+def test_deprecated_interfaces(
+    failed_plugins, successful_interfaces, successful_plugins
+):
+    """Test the "old" deprecated interfaces.
+
+    This function will be removed once all interfaces are moved to the "new" setup.
+    """
+
+    out_dicts = {}
+
     interfaces = [
-        # "stable.reader",
-        # "dev.alg",
         "dev.boundaries",
-        # "dev.cmap",
-        # "dev.filename",
         "dev.gridlines",
-        # "dev.interp",
-        # "dev.output",
-        # "dev.procflow",
         "dev.product",
     ]
 
@@ -53,21 +51,107 @@ def main():
         )
         try:
             out_dict = test_curr_interface()
+            out_dicts[curr_interface] = out_dict
         except Exception:
             print(traceback.format_exc())
             raise
 
-        print(f"SUCCESSFUL INTERFACE {curr_interface}")
-
         ppprinter = pprint.PrettyPrinter(indent=2)
         ppprinter.pprint(out_dict)
 
+    for intname in out_dicts:
+        failed_one = False
+        out_dict = out_dicts[intname]
         for modname in out_dict["validity_check"]:
             if not out_dict["validity_check"][modname]:
-                print(f"FAILED INTERFACE {curr_interface} on {modname}")
-                raise TypeError(
-                    f"Failed validity check on {modname} in interface {curr_interface}"
-                )
+                failed_plugins += [f"{intname} on {modname}"]
+                failed_one = True
+            else:
+                successful_plugins += [f"{intname} on {modname}"]
+        if not failed_one:
+            successful_interfaces += [intname]
+
+    return failed_plugins, successful_interfaces, successful_plugins
+
+
+def main():
+    """Script to test all dev and stable interfaces."""
+
+    failed_interfaces = []
+    failed_plugins = []
+    successful_interfaces = []
+    successful_plugins = []
+
+    # Test all the "old" interfaces using the original logic.
+    # (
+    #     failed_plugins,
+    #     successful_interfaces,
+    #     successful_plugins,
+    # ) = test_deprecated_interfaces(
+    #     failed_plugins, successful_interfaces, successful_plugins
+    # )
+
+    interfaces = [
+        "algorithms",
+        "colormaps",
+        "filename_formats",
+        "interpolators",
+        "output_formats",
+        "procflows",
+        "readers",
+        "title_formats",
+    ]
+
+    out_dicts = {}
+
+    for curr_interface in interfaces:
+
+        print("")
+        print(f"Testing {curr_interface}...")
+
+        test_curr_interface = getattr(
+            import_module(f"geoips.interfaces"), curr_interface
+        )
+        print(f"    from geoips.interfaces import {curr_interface}")
+
+        # Open all the interfaces (not just checking call signatures)
+        # This returns a dictionary of all sorts of stuff.
+        try:
+            out_dict = test_curr_interface.test_interface()
+            out_dicts[curr_interface] = out_dict
+        except Exception:
+            print(traceback.format_exc())
+            failed_plugins += [curr_interface]
+
+    ppprinter = pprint.PrettyPrinter(indent=2)
+
+    for intname in out_dicts:
+        ppprinter.pprint(out_dict)
+        out_dict = out_dicts[intname]
+        if out_dict["all_valid"] is True:
+            successful_interfaces += [intname]
+        else:
+            failed_interfaces += [intname]
+        for modname in out_dict["validity_check"]:
+            if not out_dict["validity_check"][modname]:
+                failed_plugins += [f"{intname} on {modname}"]
+            else:
+                successful_plugins += [f"{intname} on {modname}"]
+
+    for curr_plugin in successful_plugins:
+        print(f"SUCCESSFUL PLUGIN {curr_plugin}")
+
+    for curr_interface in successful_interfaces:
+        print(f"SUCCESSFUL INTERFACE {curr_interface}")
+
+    for curr_failed in failed_plugins:
+        print(f"FAILED PLUGIN {curr_failed}")
+
+    for curr_failed in failed_interfaces:
+        print(f"FAILED INTERFACE {curr_failed}")
+
+    if len(failed_interfaces) > 0 or len(failed_plugins) > 0:
+        raise TypeError(f"Failed validity check on plugins {failed_plugins}")
 
 
 if __name__ == "__main__":