+++ geoips/interface_modules/filename_formats/geoips_fname.py
@@ -27,7 +27,7 @@ def geoips_fname(
     area_def,
     xarray_obj,
     product_name,
-    coverage,
+    coverage=None,
     output_type="png",
     output_type_dir=None,
     product_dir=None,