+++ geoips/interfaces/procflows.py
@@ -16,7 +16,8 @@ class ProcflowsInterface(BaseInterface):
 
     name = "procflows"
     deprecated_family_attr = "procflow_type"
+    required_args = {"standard": ["fnames"]}
+    required_kwargs = {"standard": ["command_line_args"]}
 
 
 procflows = ProcflowsInterface()
-