+++ geoips/interfaces/readers.py
@@ -9,7 +9,10 @@ class ReadersInterface(BaseInterface):
     """
     name = "readers"
     deprecated_family_attr = "reader_type"
+    required_args = {"standard": ["fnames"]}
+    required_kwargs = {
+        "standard": ["metadata_only", "chans", "area_def", "self_register"]
+    }
 
 
 readers = ReadersInterface()
-