+++ geoips/interfaces/title_formats.py
@@ -1,12 +1,13 @@
 from geoips.interfaces.base import BaseInterface, BasePlugin
 
 
-class TitleFormattersInterface(BaseInterface):
+class TitleFormatsInterface(BaseInterface):
     """Interface for creating GeoIPS formatted titles."""
     name = "title_formats"
     entry_point_group = "title_formats"
     deprecated_family_attr = "title_type"
+    required_args = {"standard": []}
+    required_kwargs = {"standard": []}
 
 
-title_formats = TitleFormattersInterface()
-
+title_formats = TitleFormatsInterface()